modern gcc default to C++ versions which cause cast error; fixed
see
http://stackoverflow.com/questions/4434140/narrowing-conversions-in-c0x-is-it-just-me-or-does-this-sound-like-a-breakin

+++ irm/fastonebigheader.h
@@ -105,7 +105,7 @@ typedef __m128i v4si;
 #define v4sf_to_v4si _mm_cvttps_epi32
 
 #define v4sfl(x) ((const v4sf) { (x), (x), (x), (x) })
-#define v2dil(x) ((const v4si) { (x), (x) })
+#define v2dil(x) ((const v4si) { static_cast<long long int>(x), static_cast<long long int>(x) })
 #define v4sil(x) v2dil((((unsigned long long) (x)) << 32) | (x))
 
 typedef union { v4sf f; float array[4]; } v4sfindexer;
